+++ simapp/codec.go
@@ -5,6 +5,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/auth/vesting"
 	distr "github.com/cosmos/cosmos-sdk/x/distribution"
+	"github.com/cosmos/cosmos-sdk/x/params"
 	"github.com/cosmos/cosmos-sdk/x/slashing"
 	"github.com/cosmos/cosmos-sdk/x/staking"
 )
@@ -12,10 +13,11 @@ import (
 // AppCodec defines the application-level codec. This codec contains all the
 // required module-specific codecs that are to be provided upon initialization.
 type AppCodec struct {
-	amino    *codec.Codec
-	Slashing *slashing.Codec
+	amino *codec.Codec
 
+	Params       *params.Codec
 	Staking      *staking.Codec
+	Slashing     *slashing.Codec
 	Distribution *distr.Codec
 }
 
@@ -24,6 +26,7 @@ func NewAppCodec() *AppCodec {
 
 	return &AppCodec{
 		amino:        amino,
+		Params:       params.NewCodec(amino),
 		Staking:      staking.NewCodec(amino),
 		Slashing:     slashing.NewCodec(amino),
 		Distribution: distr.NewCodec(amino),